Avoid users being able to search appointments in the past
[FRONTEND PART]

Added an alert message when users try to schedule appointments in the past.

Not sure whether this would be good but shouldn't we remove past appointment blocks on which appointments weren't created as they are of no use anymore?

[BACKEND]
Shouldn't be hard to add a similar check in backend part as well, but could someone please point out which lambda function is used in fetching appointments?