Attach backend and frontend
I have linked edited the backend and frontend for uploading profile picture in for student. Tested once after changing the lambda functions on AWS {reverted lambda function back to original after testing to avoid breaking network in case there is some bug at some edge case}. Also, added get_student_settings.py file and edited it. This file was being used and was uploaded in the AWS but not on github.