Use fuzzywuzzy package
Fuzzywuzzy is a python package for string matching which was ported to JS as 'fuzzball'. It already had all the basic algorithm we needed.
So, instead of rewriting that, I just included it.

Also, I removed the dscaling with count at end and added it above because we were also scaling the score of StringDistance.token_set_ratio(name, supporter.name) to 1, which we already had scaled earlier. Therefore, doing that at end, we were giving stateTags and stateTopics more weight than string distance.

Furthermore, I have also scaled rating score to 1 and then given it a weight of 0.0001 (like before). We didn't need the condition (rating <= supporter.rating) because if it isn't, we will get a negative number in (supporter.rating - rating) which will bring score down and if it is greater than it will give a positive value.