Cleanup code
While working on this code for approximate string matching, I realized that there were some variables which weren't needed.
For example, instead of using a scores variable[], we could just add another property to our json named score (line 216). Also, it didn't make sense (to me) to sort an array with ids of supporters according to score and then use that to get a sortedList variable. Instead, we could just sort the newList directly (assuming there wasn't some other use for that which I couldn't think of). Also, instead of sortedList, we could reuse the same variable newList and save doubling the memory.